\name{r_HESS_SEM-package}
\alias{r_HESS_SEM-package}
\alias{r_HESS_SEM}
\docType{package}
\title{\packageTitle{r_HESS_SEM}}
\description{\packageDescription{r_HESS_SEM}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{r_HESS_SEM}
  \packageIndices{r_HESS_SEM}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{r_HESS_SEM}
  
  Maintainer: \packageMaintainer{r_HESS_SEM}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
